+++ core/src/commonMain/kotlin/pro/respawn/flowmvi/dsl/LazyPluginBuilder.kt
@@ -1,12 +1,10 @@
 package pro.respawn.flowmvi.dsl
 
-import pro.respawn.flowmvi.api.FlowMVIDSL
 import pro.respawn.flowmvi.api.LazyPlugin
 import pro.respawn.flowmvi.api.MVIAction
 import pro.respawn.flowmvi.api.MVIIntent
 import pro.respawn.flowmvi.api.MVIState
 import pro.respawn.flowmvi.api.StoreConfiguration
-import pro.respawn.flowmvi.api.StorePlugin
 
 /**
  * A class that creates a [LazyPlugin].
@@ -17,25 +15,3 @@ import pro.respawn.flowmvi.api.StorePlugin
 public class LazyPluginBuilder<S : MVIState, I : MVIIntent, A : MVIAction> @PublishedApi internal constructor(
     public val config: StoreConfiguration<S>,
 ) : StorePluginBuilder<S, I, A>()
-
-/**
- * Build a new [StorePlugin] using [StorePluginBuilder] lazily.
- * Plugin will be created upon first usage (i.e. installation).
- * @see [StorePlugin]
- */
-@FlowMVIDSL
-public inline fun <S : MVIState, I : MVIIntent, A : MVIAction> lazyPlugin(
-    @BuilderInference crossinline builder: LazyPluginBuilder<S, I, A>.() -> Unit,
-): LazyPlugin<S, I, A> = LazyPlugin {
-    LazyPluginBuilder<S, I, A>(it).apply(builder).build()
-}
-
-/**
- * Build a new [StorePlugin] using [StorePluginBuilder].
- * See [StoreBuilder.install] to install the plugin automatically.
- * @see [StorePlugin]
- */
-@FlowMVIDSL
-public inline fun <S : MVIState, I : MVIIntent, A : MVIAction> plugin(
-    @BuilderInference builder: StorePluginBuilder<S, I, A>.() -> Unit,
-): StorePlugin<S, I, A> = StorePluginBuilder<S, I, A>().apply(builder).build()
