+++ core/src/commonMain/kotlin/pro/respawn/flowmvi/logging/DefaultLoggers.kt
@@ -13,6 +13,10 @@ public val ConsoleStoreLogger: StoreLogger by lazy {
 }
 
 /**
- * A [StoreLogger] instance for each supported platform
+ * A platform-specific [StoreLogger] implementation that uses the current OS's log stream.
+ * This logger should be used as the default choice for platform-specific logging requirements.
+ *
+ * @see NoOpStoreLogger for no-op logging
+ * @see ConsoleStoreLogger for basic console output
  */
 public expect val PlatformStoreLogger: StoreLogger
