+++ core/src/jvmTest/kotlin/pro/respawn/flowmvi/test/decorator/RetryStrategyTest.kt
@@ -4,7 +4,21 @@ import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.shouldBe
 import pro.respawn.flowmvi.decorators.RetryStrategy
+import kotlin.time.Duration
 import kotlin.time.Duration.Companion.seconds
+
+// todo - test the rest of the decorator
+
+private fun testDelays(s: RetryStrategy, vararg delays: List<Duration>) {
+    repeat(s.retries) { i ->
+        // attempts start at 1 from the client
+        val attempt = i + 1
+        s.shouldRetry(attempt) shouldBe true
+        s.delay(attempt).shouldBe(delays[i], "delay for $attempt should be ${delays[i]}")
+    }
+    // 4 th retry not allowed
+    s.shouldRetry(s.retries + 1) shouldBe false
+}
 
 class RetryStrategyTest : FreeSpec({
 
@@ -25,31 +39,15 @@ class RetryStrategyTest : FreeSpec({
         "and delayInitially = false" - {
             val delayInitially = false
             "then the delays are 0, 1, 2" {
-                val delays = listOf(0.seconds, 1.seconds, 2.seconds)
                 val s = RetryStrategy.ExponentialDelay(retries, delay, exp, delayInitially)
-                repeat(retries) { i ->
-                    // attempts start at 1 from the client
-                    val attempt = i + 1
-                    s.shouldRetry(attempt) shouldBe true
-                    s.delay(attempt).shouldBe(delays[i], "delay for attempt $attempt should be ${delays[i]}")
-                }
-                // 4 th retry not allowed
-                s.shouldRetry(retries + 1) shouldBe false
+                testDelays(s, listOf(0.seconds, 1.seconds, 2.seconds))
             }
         }
         "and delayInitially = true" - {
             val delayInitially = true
             "then the delays are 1, 2, 4" {
-                val delays = listOf(1.seconds, 2.seconds, 4.seconds)
                 val s = RetryStrategy.ExponentialDelay(retries, delay, exp, delayInitially)
-                repeat(retries) { i ->
-                    // attempts start at 1 from the client
-                    val attempt = i + 1
-                    s.shouldRetry(attempt) shouldBe true
-                    s.delay(attempt).shouldBe(delays[i], "delay for $attempt should be ${delays[i]}")
-                }
-                // 4 th retry not allowed
-                s.shouldRetry(retries + 1) shouldBe false
+                testDelays(s, listOf(1.seconds, 2.seconds, 4.seconds))
             }
         }
     }
