+++ savedstate/build.gradle.kts
@@ -26,7 +26,6 @@ kotlin {
 
     sourceSets {
         val nonBrowserMain by getting
-        val browserMain by getting
         nativeMain.dependencies {
             implementation(libs.kotlin.io)
         }
